Fix error of empty list given by a dateless todo heading
Fix error of empty list given by a dateless todo heading. An empty todo heading such as a project heading has no date but the has_date() function yeilded an empty list. This lead a error where has_date()[0] does not exist so a condition of a non-empty list was added.